Changes:- 1) Removed multiple server creation for every request. 2) Added name for the module in genserver 3) Moved the code in server and from data dump server directly calling data parser server 4) Removed extract data from data parser and moved it to test since it is being used in it only.
