+++ apps/acqdat_api/lib/acqdat_api_web/controllers/role-management/user_controller.ex
@@ -45,7 +45,9 @@ defmodule AcqdatApiWeb.RoleManagement.UserController do
 
         with {:extract, {:ok, data}} <- {:extract, extract_changeset_data(changeset)},
              {:create, {:ok, user}} <- {:create, User.create(data)} do
-          Task.start_link(ElasticSearch.create_user("organisation", user, %{id: user.org_id}))
+          Task.start_link(fn ->
+            ElasticSearch.create_user("organisation", user, %{id: user.org_id})
+          end)
 
           conn
           |> put_status(200)
